fix(tracing): satisfy mypy, import-linter, and SavedInsights metadata
- aggregation_query_runner.py: cast `prop` to `SpanPropertyFilter` when
  routing into the typed lists (mypy can't narrow from getattr); use
  `self.query.kind` directly (already `Literal[str]`, no `.value`);
  declare `modifiers/timings/limit_context` on the mixin with the same
  types `QueryRunner` uses so the multi-inheritance MRO type-checks.
- Move `run_aggregation_query` / `run_tree_query` from
  `aggregation_query_runner.py` into `logic.py` so views.py imports
  from the existing facade-grandfathered module (import-linter rule
  `presentation must use facade`).
- SavedInsights.tsx: add `Record<NodeKind, InsightTypeMetadata>` entries
  for the two new node kinds. Required by the exhaustiveness check.

Generated-By: PostHog Code
Task-Id: f029c67a-c362-4104-8c2d-be2f440f147e
